(1) Added logic and validation to delete time off (2) Delete is now a detail view which includes delete...

+++ implementation/app/business/timetracking/TimeOffServiceImpl.java
@@ -3,12 +3,15 @@ package business.timetracking;
 import business.notification.NotificationSender;
 import business.notification.NotificationType;
 import business.usermanagement.InternalUserManagement;
+import business.usermanagement.NotAuthorizedException;
+import business.usermanagement.UserException;
 import com.google.inject.Inject;
 import infrastructure.TimeOffRepository;
 import models.Notification;
 import models.TimeOff;
 import models.User;
 import org.joda.time.DateTime;
+import org.joda.time.DateTimeComparator;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -219,15 +222,20 @@ class TimeOffServiceImpl implements TimeOffService {
     @Override
     public void deleteTimeTrack(int userId, int id) throws Exception {
         TimeOff timeOffToDelete = _repository.readTimeOff(id);
+
         if(timeOffToDelete.getUser().getId() != userId){
-            throw new Exception("Not allwoed");
+            throw new NotAuthorizedException("User " + userId + " is not allowed to delete time off " + id);
+        }
+
+        if(timeOffToDelete.getFrom().isBeforeNow()){
+            throw new UserException("exceptions.timeoff.error_delete_timoff_in_past");
         }
 
         User employee = timeOffToDelete.getUser();
         User boss = employee.get_boss();
-        Notification ntification = new Notification(NotificationType.INFORMATION, "Deleted time off", employee, boss);
+        Notification notification = new Notification(NotificationType.INFORMATION, "Deleted time off", employee, boss);
 
-        _notificationSender.sendNotification(ntification);
+        _notificationSender.sendNotification(notification);
         _repository.deleteTimeOff(timeOffToDelete);
     }
 }
